db changes

+++ src/database/models/citizen.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Date
+from sqlalchemy import Column, Integer, String, Date, BigInteger
 from src.database.base_model import BaseModel
 
 class Citizen(BaseModel):
@@ -9,7 +9,7 @@ class Citizen(BaseModel):
     name        = Column(String(50))
     surname     = Column(String(100))
     birth_date  = Column(Date)
-    pesel       = Column(Integer, unique=True)
+    pesel       = Column(BigInteger, unique=True)
 
     def __repr__(self):
         return "<Citizen: id = %d  name = %s  surname = %s  birth_date = %s  pesel = %d>" % (
